requirements.txt
Django>=4.2
psycopg2-binary
django-ckeditor
Pillow>=10.0
requests>=2.31.0
python-dotenv>=1.0.0